+++ lib/flutter_joystick.dart
@@ -3,4 +3,6 @@ library flutter_joystick;
 export 'src/joystick.dart';
 export 'src/joystick_area.dart';
 export 'src/joystick_base.dart';
+export 'src/joystick_controller.dart';
+export 'src/joystick_stick.dart';
 export 'src/joystick_stick_offset_calculator.dart';