+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.4
+
+* Added: Support for customizing `drawArrows`, `size`, and `color` parameters in `JoystickBase`.
+* Added: Support for customizing `size` parameter in `JoystickStick`.
+
 ## 0.0.3
 
 * Added a send of zero offset to the joystick listener when the stick is released